+++ mobile/android/geckoview/src/androidTest/assets/web_extensions/test-support/background.js
@@ -80,6 +80,9 @@ const APIS = {
   IsSessionHistoryInParentRunning() {
     return browser.test.isSessionHistoryInParentRunning();
   },
+  IsFissionRunning() {
+    return browser.test.isFissionRunning();
+  },
   TriggerCookieBannerDetected({ tab }) {
     return browser.test.triggerCookieBannerDetected(tab.id);
   },