+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/rule/GeckoSessionTestRule.java
@@ -2800,6 +2800,11 @@ public class GeckoSessionTestRule implements TestRule {
     return (Boolean) webExtensionApiCall("IsSessionHistoryInParentRunning", null);
   }
 
+  /** Checks if fission is running. */
+  public boolean isFissionRunning() {
+    return (Boolean) webExtensionApiCall("IsFissionRunning", null);
+  }
+
   private Object webExtensionApiCall(
       final @NonNull String apiName, final @NonNull SetArgs argsSetter) {
     return webExtensionApiCall(null, apiName, argsSetter);